Move runner node utility functions into runnerNode class
