+++ mlir/include/air/Util/Util.h
@@ -38,6 +38,13 @@ scf::ParallelOp getParallelRegionInitValsOwner(Operation *op, Value val);
 HerdOp getHerdArgOwner(Value val);
 // Get the parent air.hierarchy op of a tile id
 HierarchyInterface getHierarchyArgOwner(Value val);
+// Get the scf parent op from scf.yield op
+template <typename T> Operation *getScfParentOpFromYieldOp(scf::YieldOp op) {
+  if (auto scfop = dyn_cast<T>(op->getParentOp())) {
+    return scfop.getOperation();
+  }
+  return nullptr;
+}
 
 // Erase a kernel operand from air.hierarchy op
 void eraseAIRHierarchyOperand(HierarchyInterface op, unsigned index);
